Another clean up upgrade & a Catalog Items index() upgrade.
You can now name the category parameter for the index to show only a particular category of items in the catalog_items table.

This makes me think categories should show an index() view file instead of trying to get the categories view to show the right view for each list of category items, we should instead rely on the model that is related to control what the view for the category looks like.