Added Sorting Element.
Using the paginator sorting links there is a new element which can display a sorting select form, which will allow you to sort the contents of an index page by selecting the sort fields from the drop down.

As of now the only use is on the catalog_items/index page for testing.

To add to new pages or just to set the variables for it in the view file is very similar to the context_menu.  Use the catalog items index view file for an example.