Plugin Install Upgrade.
You can now install plugins that aren't already installed by going to /install for any site.  On that page you will see a list of plugins that are available to install.

This means that we have to keep every Plugin Config schema.php file up to date now.   This also means that we are close to getting rid of the version directory all together.  The missing piece is a script that checks the date header of the schema files to see if all of the plugins are up to date.   Then it is a matter of running the command : cake schema generate --plugin SomePlugin after you make a database upgrade.

IMPORTANT NOTE : The cake console only works with the app/Config/database file, not individual site files.  But the plugin install does work with individual sites.